python/chess_to_script/requirements.txt
cairosvg
chess
google-cloud-firestore
google-cloud-pubsub
google-cloud-storage
google-cloud-texttospeech
pyfarmhash